X-9 Follow-up to X-7 TreeMap implementation
