+++ src/items/rfm-customer/components/TreeMapHoveredLayer.tsx
@@ -1,6 +1,6 @@
 import type { CustomLayerProps } from '@nivo/treemap';
 
-import { RFMSegmentIds } from '@/stories/rfm-elasticity/types/RFMSegmentId';
+import { RFMSegmentIds } from '@/items/rfm-elasticity/types/RFMSegmentId';
 
 import type { TreeMapNode } from '../types/SegmentTreeMapData';
