+++ src/items/rfm-customer/hooks/useJumpToSegmentElasticity.ts
@@ -2,7 +2,7 @@ import { useCallback } from 'react';
 
 import { useActiveDashboard, useRFMSegment } from '@/app/providers';
 import { DashboardId } from '@/common/types/DashboardIds';
-import type { RFMSegmentIds } from '@/stories/rfm-elasticity/types/RFMSegmentId';
+import type { RFMSegmentIds } from '@/items/rfm-elasticity/types/RFMSegmentId';
 
 export function useJumpToSegmentElasticity() {
   const { setActiveDashboard } = useActiveDashboard();
