+++ src/items/rfm-elasticity/components/RFMKPISection.tsx
