+++ src/items/rfm-elasticity/model/RFMSegments.ts
