+++ src/items/rfm-elasticity/types/BaseAccordionItem.ts
