+++ src/items/rfm-elasticity/types/RFMSegmentKPIData.ts
