+++ src/items/sidebar/components/SidebarSimulationGroup.tsx
